+++ doc/changelog.rst
@@ -824,12 +824,26 @@ process too much.It could be used for archiving though.
 Miscellaneous
 -------------
 
-* smooothing_deg option in
+* smoothing_deg option in
   :py:meth:`orb.utils.fft.transform_interferogram` has 1 sample less
   than replaced by a more robust smoothing_coeff option. The smoohting
   degree is now defined as smoothing_coeff * interferogram_size. In a
   general way the smoothing degree (the number of samples smoothed at
   a transition between a part of zeros and a part of signal in the
   interferogram) is now bigger because -25/+100 interferogram present
-  a very sharp transition on the left side whcih is very near ZPD and
+  a very sharp transition on the left side which is very near ZPD and
   create large wiggles in the spectra. A higher smoothing degree is the solution.
+
+
+v.2.0.1
+=======
+
+Fit module
+----------
+
+* :py:meth:`~cutils.sincgauss1d`: sigma minimum limit set at
+  1e-5. Below this value a pure sinc is returned.
+
+* small bug correction
+
+